backend/requirements.txt
fastapi[standard]
pydantic
sqlalchemy
passlib
bcrypt==3.2.2
python-jose
websockets
